Refactor codes

+++ mymalloc.h
@@ -1,5 +1,3 @@
-#include <stdio.h>
-
 /*****************************************************************************
 
 mymalloc.h
@@ -9,11 +7,13 @@ Also, initializes struct.
  
 ******************************************************************************/
 
+#include <stdio.h>
+
 /* Struct to keep track of size of each block, next block and whether it's free or not */
 struct block{
- int size;
- int free;
- struct block *next;
+	size_t size;
+	int free;
+	struct block *next;
 }block;
 
 /* Global structures free_list, current and previous */
@@ -23,7 +23,7 @@ struct block *previous;
 
 /* Function declarations */
 void initialize_freelist();
-void split_block(int size);
-void *my_malloc(int no_of_bytes);
+void split_block(size_t size);
+void *my_malloc(size_t no_of_bytes);
 void merge_blocks();
 void my_free(void* ptr);
